Use labels for recipes instead of ingredients
