+++ app/controllers/forms/recipes.py
@@ -4,7 +4,7 @@ from wtforms.widgets import TextArea
 
 from flask_wtf import FlaskForm
 
-from wtforms_sqlalchemy.fields import QuerySelectField
+from wtforms_sqlalchemy.fields import QuerySelectField, QuerySelectMultipleField
 
 
 def categories():
@@ -19,6 +19,18 @@ def measurements():
     return Measurement.load_all()
 
 
+def labels():
+    from app.models.labels import Label
+
+    return Label.load_all()
+
+
+def dietary_labels():
+    from app.models.labels import Label
+
+    return Label.load_dietary()
+
+
 class RecipesForm(FlaskForm):
     name = StringField(
         "Název receptu", [validators.InputRequired("Název musí být vyplněn")]
@@ -28,6 +40,8 @@ class RecipesForm(FlaskForm):
 
     category = QuerySelectField("Kategorie", query_factory=categories, allow_blank=True)
     portion_count = IntegerField("Počet porcí", [validators.NumberRange(min=1)])
+
+    labels = QuerySelectMultipleField("Dietní omezení", query_factory=dietary_labels)
 
     submit = SubmitField("Přidat recept")
