+++ app/templates/ingredients/edit.html.j2
@@ -30,8 +30,6 @@
 
             <hr>
 
-            {% include "ingredients/_form_labels.html.j2" %}
-
             {% include "ingredients/_form_nutrition_values.html.j2" %}
             
             <button type="submit" class="btn btn-warning">{{ texts.edit_confirm }} {{ icon("edit") }}</button>
