+++ app/templates/ingredients/new.html.j2
@@ -16,8 +16,6 @@
     {{ render_field(form.description) }}
     {{ render_field(form.measurement) }}
     {{ render_field(form.category) }}
-    
-    {% include "ingredients/_form_labels.html.j2" %}
 
     {% include "ingredients/_form_nutrition_values.html.j2" %}
